// src/app/models/anamnese/PutAnamneseModel.ts
export class PutAnamneseModel{
  
    idAnamnese: string = '';
    dia_hora: string = '';
   empresa: string = '';
   nome: string = '';
   telefone: string = '';
   data_admissao: string = '';
   data_nascimento: string = '';
   cpf: string = '';
   rg: string = '';
   orgaoemissor: string = '';
   setor: string = '';
   funcao: string = '';
   tipo_aso: string = '';
   tubercolose: string = '';
   pressaoalta: string = '';
   alcoolismo: string = '';
   cancer: string = '';
   diabetes: string = '';
   infarto: string = '';
   doencao_nervosa: string = '';
   doenca_da_tireoide: string = '';
   alergias: string = '';
   bebidas_alcoolicas: string = '';
   atividade_fisica: string = '';
   tabagismo: string = '';
   transporte: string = '';
   tempo_deslocamento: string = '';
   leitura: string = '';
   cinema_teatro: string = '';
   videogame: string = '';
   isntrumentos: string = '';
   canto: string = '';
   tv: string = '';
   basquete_volei: string = '';
   modelagem: string = '';
   musica_com_head_fone: string = '';
   musica_sem_head_fone: string = '';
   futebol: string = '';
   lutas: string = '';
   musculacao: string = '';
   jardinagem: string = '';
   tenis: string = '';
   jogos_computador: string = '';
   navegar_internet: string = '';
   outros: string = '';
   historia_ocup1: string = '';
   historia_ocup2: string = '';
   historia_ocup3: string = '';
   historia_ocup4: string = '';
   historia_ocup5: string = '';
   historia_ocup6: string = '';
   oculos: string = '';
   corrigir_miopia: string = '';
   corrigir_antigmatismo: string = '';
   corrigir_hipermetropia: string = '';
   corrigir_vista_cansada: string = '';
   corrigir_nao_sei: string = '';
   cistite: string = '';
   diabete: string = '';
   tonteiras: string = '';
   renite_sinusite: string = '';
   varises: string = '';
   dor_de_cabeca: string = '';
   prisao_ventre: string = '';
   hemorraidas: string = '';
   pressao_alta: string = '';
   hernia: string = '';
   dor_no_peito: string = '';
   dor_nas_pernas: string = '';
   resfriados: string = '';
   problemas_tireoides: string = '';
   epilepsia_convulcoes: string = '';
   asma_bronquite: string = '';
   tubercolose1: string = '';
   ulcera: string = '';
   tosse_cronica: string = '';
   pedra_nos_rins: string = '';
   remedios: string = '';
   hepatite_a: string = '';
   hepatite_b: string = '';
   hepatite_c: string = '';
   hepatite_nao_sei: string = '';
   operacao: string = '';
   tipo_operacao: string = '';
   ano_operacao: string = '';
   quebra_osso: string = '';
   quebra_osso_onde: string = '';
   dor_dedo_direito: string = '';
   dor_dedo_esquerdo: string = '';
   dor_maos_direito: string = '';
   dor_maos_esquerdo: string = '';
   dor_punho_direito: string = '';
   dor_punho_esquerdo: string = '';
   dor_cotovelo_direito: string = '';
   dor_cotovelo_esquerdo: string = '';
   dor_braco_direito: string = '';
   dor_braco_esquerdo: string = '';
   dor_ombro_direito: string = '';
   dor_ombro_esquerdo: string = '';
   dor_joelho_direito: string = '';
   dor_joelho_esquerdo: string = '';
   dor_pescoco: string = '';
   dor_costas_superior: string = '';
   dor_costas_medias: string = '';
   dor_costas_inferior: string = '';
   doenca_sexualmente_transmissivel: string = '';
   prostata: string = '';
   anticosepsional: string = '';
   aborto: string = '';
   preventivo_no_ano: string = '';
   data_mestruacao: string = '';
   tensao_mestruacao: string = '';
   status_mestruacao: string = '';
   colicas: string = '';
   gravidez: string = '';
   pressao_arterial: string = '';
   frequencia_cardiaca: string = '';
   batimentos_por_minuto: string = '';
   icm: string = '';
   altura: string = '';
   peso: string = '';
   ganglios: string = '';
   dentes: string = '';
   tireoide: string = '';
   orofaringe: string = '';
   acuidade_visual: string = '';
   ritimo_cardiaco: string = '';
   ritimo_cardiaco_obs: string = '';
   sopros: string = '';
   pulmoes: string = '';
   abdomen: string = '';
   maos_punhos: string = '';
   perda_forca_de_preensao: string = '';
   tinel: string = '';
   phalen: string = '';
   filkenstein: string = '';
   cotovelo_direito: string = '';
   cotovelo_esquerdo: string = '';
   ombro_direito: string = '';
   ombro_esquerdo: string = '';
   coluna: string = '';
   coluna_outros: string = '';
   aparelho_muscular: string = '';
   membros_inferiores: string = '';
   membros_inferiores_obs: string = '';
   joelho_direito: string = '';
   joelho_esquerdo: string = '';
   exames_complementares: string = '';
   risco_quimico: string = '';
   risco_fisico: string = '';
   risco_biologico: string = '';
   risco_ergonomico: string = '';
   risco_especifico: string = '';
   exame_complementar: string = '';
   parecer_especialista: string = '';
   encaminhamento_perito: string = '';
   cat: string = '';
   observacoes: string = '';
   status_anamnese: string = '';
   

   
}